backend/requirements.txt
fastapi
uvicorn
python-dotenv
boto3
python-multipart
python-dotenv

